--- src/features/auth/schema.ts
@@ -1,8 +1,0 @@
-import z from 'zod';
-
-export const authSchema = z.object({
-  username: z.string(),
-  password: z.string(),
-});
-
-export type AuthSchema = z.infer<typeof authSchema>;
