--- src/features/counter/Counter.tsx
@@ -1,25 +1,0 @@
-import { decrement, increment } from './counterSlice';
-import { useAppDispatch } from '@/app/hooks';
-
-export const Counter = () => {
-  const dispatch = useAppDispatch();
-
-  return (
-    <div>
-      <div>
-        <button
-          aria-label="Increment value"
-          onClick={() => dispatch(increment())}
-        >
-          Increment
-        </button>
-        <button
-          aria-label="Decrement value"
-          onClick={() => dispatch(decrement())}
-        >
-          Decrement
-        </button>
-      </div>
-    </div>
-  );
-};
