+++ src/main.tsx
@@ -28,9 +28,9 @@ const router = createRouter({
 
 // Register the router instance for type safety
 declare module '@tanstack/react-router' {
-  interface Register {
+  type Register = {
     router: typeof router;
-  }
+  };
 }
 
 const darkTheme = createTheme({
