+++ src/queries/customer.queries.ts
@@ -2,7 +2,7 @@ import { createQueryKeys } from '@lukemorales/query-key-factory';
 import { axiosBase } from '@/plugins/axiosBase';
 
 import { PATH } from '@/constant/routes';
-import type { ResponseCustomerListSchema, CustomerDetailSchema } from '@/features/customer/schema';
+import type { ResponseCustomerListSchema, CustomerDetailSchema } from '@/routes/customer/-types/schema';
 
 export const CustomerQueries = createQueryKeys('customer', {
   getCustomerList: () => ({
@@ -18,7 +18,7 @@ export const CustomerDetailQueries = createQueryKeys('customer', {
   getCustomerDetail: (id: number) => ({
     queryKey: ['getCustomerDetail'],
     queryFn: () =>
-      axiosBase<CustomerDetailSchema>(PATH.CUSTOMER + `/${id}`).then((res) => {
+      axiosBase<CustomerDetailSchema>(PATH.CUSTOMER_DETAIL.replace('$id', String(id))).then((res) => {
         return res.data;
       }),
   }),
