+++ src/routes/-components/Top.tsx
@@ -5,7 +5,7 @@ import { useAppSelector } from '@/app/hooks';
 import { PATH } from '@/constant/routes';
 import { useAuthCheck } from '@/hooks/useAuthCheck';
 import { SimpleButton } from '@/components/atoms/SimpleButton';
-import { SimpleCard } from '@/components/molecules/SimpleCard';
+import { SimpleCard } from './SimpleCard';
 
 import { ButtonStyle, CardStyle } from './Top.css';
 
@@ -25,13 +25,6 @@ export const Top = () => {
 
       <SimpleCard minWidth={275} countStore={countStore} />
 
-      <div className={ButtonStyle}>
-        <SimpleButton
-          buttonName='Page to Login'
-          color='secondary'
-          onClick={() => navigate({ to: PATH.LOGIN })}
-        ></SimpleButton>
-      </div>
       <div className={ButtonStyle}>
         <SimpleButton buttonName='Page to Register' onClick={() => navigate({ to: PATH.REGISTER })}></SimpleButton>
       </div>
