+++ src/routes/customer/-components/Customer.tsx
@@ -11,7 +11,6 @@ export const Customer = () => {
   const navigate = useNavigate();
 
   const location = useLocation();
-  console.log(location.pathname);
 
   return (
     <>
