+++ src/routes/login/route.lazy.tsx
@@ -1,0 +1,10 @@
+import { createLazyFileRoute } from '@tanstack/react-router';
+import { Login } from './-components/Login';
+
+export const Route = createLazyFileRoute('/login')({
+  component: LoginPage,
+});
+
+function LoginPage() {
+  return <Login />;
+}
